Refactor command registration functions into our argparse library
The former existed before the latter, but now that the latter exists,
it's a more appropriate place to consolidate them.
